habiapi/requirements.txt
attrs==21.4.0
iniconfig==1.1.1
mysql-connector-python==8.0.28
packaging==21.3
pluggy==1.0.0
protobuf==3.20.0
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.8
pytest==7.1.1
tomli==2.0.1
typing-extensions==4.1.1
